+++ packages/formik/src/Formik.tsx
@@ -48,6 +48,10 @@ type FormikMessage<Values> =
   | { type: 'SET_FIELD_ERROR'; payload: { field: string; value?: string } }
   | { type: 'SET_TOUCHED'; payload: FormikTouched<Values> }
   | { type: 'SET_ERRORS'; payload: FormikErrors<Values> }
+  | {
+      type: 'SET_LOW_PRIORITY_ERRORS';
+      payload: { values: Values; errors: FormikErrors<Values> };
+    }
   | { type: 'SET_STATUS'; payload: any }
   | {
       type: 'SET_FORMIK_STATE';
@@ -84,6 +88,23 @@ function formikReducer<Values>(
       }
 
       return { ...state, errors: msg.payload };
+    case 'SET_LOW_PRIORITY_ERRORS':
+      if (
+        // Low priority validation can occur after high priority validation and
+        // this will create stale validation results, e.g:
+        // SET_FIELD_VALUE-------------------SET_LOW_PRIORITY_ERRORS
+        //   SET_FIELD_VALUE-------------------SET_LOW_PRIORITY_ERRORS
+        //       SET_ISVALIDATING-SET_ERRORS-SET_ISVALIDATING
+        //
+        // So we want to skip validation results if values are not the same
+        // anymore.
+        isEqual(state.errors, msg.payload.errors) ||
+        !isEqual(state.values, msg.payload.values)
+      ) {
+        return state;
+      }
+
+      return { ...state, errors: msg.payload.errors };
     case 'SET_STATUS':
       return { ...state, status: msg.payload };
     case 'SET_ISSUBMITTING':
@@ -348,7 +369,10 @@ export function useFormik<Values extends FormikValues = FormikValues>({
         return runAllValidations(values)
           .then(combinedErrors => {
             if (!!isMounted.current) {
-              dispatch({ type: 'SET_ERRORS', payload: combinedErrors });
+              dispatch({
+                type: 'SET_LOW_PRIORITY_ERRORS',
+                payload: { values, errors: combinedErrors },
+              });
             }
             return combinedErrors;
           })