+++ packages/formik-native/CHANGELOG.md
@@ -14,6 +14,15 @@
 - Updated dependencies [[`5efd691`](https://github.com/formium/formik/commit/5efd691b8784fda6645d362189f55c618f030758)]:
   - formik@3.0.0-next.0
 
+## 2.1.10
+
+### Patch Changes
+
+- [`00f95ec`](https://github.com/formium/formik/commit/00f95ec4ec5266eed8ad4e97b76321205c704d51) [#2854](https://github.com/formium/formik/pull/2854) Thanks [@umidbekkarimov](https://github.com/umidbekkarimov)! - Fix low priority validation race condition.
+
+- Updated dependencies [[`00f95ec`](https://github.com/formium/formik/commit/00f95ec4ec5266eed8ad4e97b76321205c704d51)]:
+  - formik@2.2.2
+
 ## 2.1.9
 
 ### Patch Changes